requirements.txt
python-dotenv==1.0.1
requests==2.32.1
selenium==4.21.0